Refactoring to HomeActivity and WiFiDebugActivity
