+++ app/src/androidTest/java/ch/tarsier/tarsier/test/HomeActivityTest.java
@@ -6,19 +6,22 @@ import ch.tarsier.tarsier.HomeActivity;
 import ch.tarsier.tarsier.R;
 
 import static com.google.android.apps.common.testing.ui.espresso.Espresso.onView;
+import static com.google.android.apps.common.testing.ui.espresso.Espresso.pressBack;
+import static com.google.android.apps.common.testing.ui.espresso.action.ViewActions.clearText;
 import static com.google.android.apps.common.testing.ui.espresso.action.ViewActions.click;
+import static com.google.android.apps.common.testing.ui.espresso.action.ViewActions.typeText;
 import static com.google.android.apps.common.testing.ui.espresso.assertion.ViewAssertions.matches;
+import static com.google.android.apps.common.testing.ui.espresso.matcher.ViewMatchers.isClickable;
 import static com.google.android.apps.common.testing.ui.espresso.matcher.ViewMatchers.withId;
-import static com.google.android.apps.common.testing.ui.espresso.matcher.ViewMatchers.withText;
 
 /**
- * Created by benjamin on 22/10/14.
+ * Created by Benjamin Paccaud on 22/10/14.
  */
 public class HomeActivityTest extends ActivityInstrumentationTestCase2<HomeActivity> {
 
     @SuppressWarnings("deprecation")
     public HomeActivityTest() {
-        super("ch.tarsier.tarsier",HomeActivity.class);
+        super("ch.tarsier.tarsier", HomeActivity.class);
     }
 
     @Override
@@ -27,9 +30,46 @@ public class HomeActivityTest extends ActivityInstrumentationTestCase2<HomeActiv
         getActivity();
     }
 
-    public void testTest() {
-        onView(withId(R.id.helloWorld)).check(matches(withText("Hello world!")));
-        onView(withId(R.id.testUIButton)).perform(click());
-        onView(withId(R.id.helloWorld)).check(matches(withText("Hello World!")));
+    public void testStartButtonClickability() {
+        onView(withId(R.id.lets_chat)).perform(click());
+        //nothing should happen
+        onView(withId(R.id.username)).perform(typeText("benpac"));
+        onView(withId(R.id.lets_chat)).perform(click());
+        //nothing should happen
+        onView(withId(R.id.status_message)).perform(typeText("my status"));
+        onView(withId(R.id.lets_chat)).check(matches(isClickable()));
+        onView(withId(R.id.username)).perform(clearText());
+        onView(withId(R.id.lets_chat)).perform(click());
+        //nothing should happen
+        onView(withId(R.id.username)).perform(typeText("Benpac"));
+        onView(withId(R.id.lets_chat)).check(matches(isClickable()));
+        onView(withId(R.id.lets_chat)).perform(click());
+        //we go to new activity
+        //basic check
+        //pressBack();
+        //we are back to home activity
     }
+
+    public void testAddPictureClick() {
+        onView(withId(R.id.picture)).check(matches(isClickable()));
+        onView(withId(R.id.picture)).perform(click());
+        // we are on AddProfilePictureActivity
+        onView(withId(R.id.add_from_existing)).check(matches(isClickable()));
+        onView(withId(R.id.add_new_picture)).check(matches(isClickable()));
+        //click and check intent for existing picture
+        onView(withId(R.id.add_from_existing)).perform(click());
+        //get from gallery
+        //pressBack();
+        //click and check intent for new picture
+        onView(withId(R.id.add_from_existing)).perform(click());
+        //get from camera
+        //pressBack();
+        //back to Home
+        pressBack();
+    }
+
+    public void testValidUsername() {
+        //TODO
+    }
+
 }
