+++ app/src/main/java/ch/tarsier/tarsier/AddProfilePictureActivity.java
@@ -1,0 +1,51 @@
+package ch.tarsier.tarsier;
+
+import android.app.Activity;
+import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.view.View;
+import android.widget.Toast;
+
+/**
+ * Activity to upload a profile picture, either from the Gallery or
+ * by taking a new picture.
+ * @author Benjamin Paccaud
+ *
+ */
+public class AddProfilePictureActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_add_profile_picture);
+    }
+
+    public void onClickNewPicture(View view) {
+        Toast.makeText(this, "new pic", Toast.LENGTH_SHORT).show();
+    }
+
+    public void onClickAddExisting(View view) {
+        Toast.makeText(this, "existing pic", Toast.LENGTH_SHORT).show();
+    }
+
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.add_profile_picture, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+        if (id == R.id.action_settings) {
+            return true;
+        }
+        return super.onOptionsItemSelected(item);
+    }
+}
