+++ app/src/main/java/ch/tarsier/tarsier/HomeActivity.java
@@ -1,167 +1,81 @@
 package ch.tarsier.tarsier;
 
 import android.app.Activity;
-import android.app.Fragment;
-import android.content.Context;
-import android.content.IntentFilter;
-import android.net.wifi.WpsInfo;
-import android.net.wifi.p2p.WifiP2pConfig;
-import android.net.wifi.p2p.WifiP2pDevice;
-import android.net.wifi.p2p.WifiP2pDeviceList;
-import android.net.wifi.p2p.WifiP2pInfo;
-import android.net.wifi.p2p.WifiP2pManager;
+import android.content.Intent;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.Message;
-import android.util.Log;
+import android.text.Editable;
+import android.text.TextWatcher;
+import android.view.Menu;
 import android.view.MenuItem;
-import java.io.IOException;
-import java.util.ArrayList;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.Toast;
+
+/**
+ * This is the Home screen of Tarsier. It allows to enter a Username
+ * and initiate a session
+ * @author Benjamin Paccaud.
+ */
+public class HomeActivity extends Activity {
+
+    static private final int MIN_USERNAME_LENGTH = 1;
+    static private final int MIN_STATUS_LENGTH = 1;
 
 
-public class HomeActivity extends Activity implements WifiP2pManager.ConnectionInfoListener,
-        WiFiDirectGroupList.DeviceClickListener,Server.MessageTarget,Handler.Callback{
-
-    public static final String TAG = "HomeActivity";
-    public static final int SERVER_PORT = 8888;
-    public static final int MESSAGE_READ = 0x400 + 1;
-    public static final int MY_HANDLE = 0x400 + 2;
-
-
-    private WifiP2pManager.PeerListListener peerListListener;
-    private IntentFilter mIntentFilter;
-    private WifiP2pManager mManager;
-    private WifiP2pManager.Channel mChannel;
-    private WiFiDirectBroadcastReceiver mReceiver;
-
-
-    private WiFiDirectGroupList groupList;
-    private final ArrayList<WifiP2pDevice> peers = new ArrayList<WifiP2pDevice>();
-    private ChatRoom chatRoom;
-    private Handler mHandler = new Handler(this);
-
-
-    public boolean testUI = true;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_home);
-
-        // Add intents that broadcast receiver checks for
-        mIntentFilter = new IntentFilter();
-        mIntentFilter.addAction(WifiP2pManager.WIFI_P2P_STATE_CHANGED_ACTION);
-        mIntentFilter.addAction(WifiP2pManager.WIFI_P2P_PEERS_CHANGED_ACTION);
-        mIntentFilter.addAction(WifiP2pManager.WIFI_P2P_CONNECTION_CHANGED_ACTION);
-        mIntentFilter.addAction(WifiP2pManager.WIFI_P2P_THIS_DEVICE_CHANGED_ACTION);
-
-
-        mManager = (WifiP2pManager) getSystemService(Context.WIFI_P2P_SERVICE);
-        mChannel = mManager.initialize(this, getMainLooper(), null);
-        mReceiver = new WiFiDirectBroadcastReceiver(mManager, mChannel, this,peerListListener);//Be careful, using peerList before initializing it
-        createPeerListener();
-        initiatePeerDiscovery();
-
-
-        mManager.requestPeers(mChannel,peerListListener);
-        groupList = new WiFiDirectGroupList();
-        getFragmentManager().beginTransaction()
-                .add(R.id.container, groupList, "groups").commit();
+        enableStartButton(false);
+        ((EditText) findViewById(R.id.username)).addTextChangedListener(new EditTextWatcher());
+        ((EditText) findViewById(R.id.status_message)).addTextChangedListener(new EditTextWatcher());
 
     }
 
-    private void createPeerListener() {
-
-        peerListListener = new WifiP2pManager.PeerListListener() {
-            @Override
-            public void onPeersAvailable(WifiP2pDeviceList peersList) {
-                peers.clear();
-                peers.addAll(peersList.getDeviceList());
-
-                // If an AdapterView is backed by this data, notify it
-                // of the change.  For instance, if you have a ListView of available
-                // peers, trigger an update.
-                WiFiDirectGroupList fragment = (WiFiDirectGroupList) getFragmentManager()
-                        .findFragmentByTag("groups");
-                if (fragment != null) {
-                    WiFiDirectGroupList.WiFiDevicesAdapter adapter = ((WiFiDirectGroupList.WiFiDevicesAdapter) fragment
-                            .getListAdapter());
-                    adapter.clear();
-                    adapter.addAll(peers);
-                    adapter.notifyDataSetChanged();
-                    Log.d(TAG, "Peer list updated: "+ peers.toString());
-                    if (peers.size() == 0) {
-                        Log.d(TAG, "No devices found");
-                        return;
-                    }
-
-                }
-            }
-        };
+    /**
+     * Toggle the clickable property of the lets_chat Button
+     * @param enable true makes the Button clickable.
+     */
+    private void enableStartButton(boolean enable) {
+        Button start = (Button) findViewById(R.id.lets_chat);
+        start.setClickable(enable);
     }
 
-    @Override
-    protected void onRestart() {
-        Fragment frag = getFragmentManager().findFragmentByTag("groups");
-        if (frag != null) {
-            getFragmentManager().beginTransaction().remove(frag).commit();
+
+    public void onClickLetsChat(View view) {
+        //create intent with username and launch the list of rooms
+        EditText username = (EditText) findViewById(R.id.username);
+        if (isValidUsername(username.getText().toString())) {
+            //continue
+        } else {
+            //show toast with information on invalidity
         }
-        super.onRestart();
+        //debug
+        Toast.makeText(this, "enabled", Toast.LENGTH_SHORT).show();
     }
 
+    private boolean isValidUsername(String username) {
+        //check on the string (length)
+        return true;
+    }
 
-    public void initiatePeerDiscovery(){
-        mManager.discoverPeers(mChannel,new WifiP2pManager.ActionListener() {
-            @Override
-            public void onSuccess() {
-                Log.d(TAG,"Peer discovery initiation succeeded");
-            }
-
-            @Override
-            public void onFailure(int reasonCode) {
-                if (reasonCode == WifiP2pManager.P2P_UNSUPPORTED) {
-                    Log.d(TAG, "Peer discovery initiation failed. P2P isn't supported on this device.");
-                }else if(reasonCode == WifiP2pManager.BUSY){
-                    Log.d(TAG, "Peer discovery initiation failed. The system is too busy to process the request.");
-                }else if(reasonCode == WifiP2pManager.ERROR){
-                    Log.d(TAG, "Peer discovery initiation failed. The operation failed due to an internal error.");
-                }
-            }
-        });
-
-
+    public void onClickAddPicture(View view) {
+        //launch Activity to have the user choose the picture
+        Intent picture = new Intent(this, AddProfilePictureActivity.class);
+        this.startActivity(picture);
+        //debug
+        //Toast.makeText(this, "add picture", Toast.LENGTH_SHORT).show();
     }
 
 
     @Override
-    public void connectP2p(WifiP2pDevice device) {
-        WifiP2pConfig config = new WifiP2pConfig();
-        config.deviceAddress = device.deviceAddress;
-        config.wps.setup = WpsInfo.PBC;
-        mManager.connect(mChannel, config, new WifiP2pManager.ActionListener() {
-            @Override
-            public void onSuccess() {
-                Log.d(TAG,"Connecting to peer");
-            }
-            @Override
-            public void onFailure(int errorCode) {
-                Log.d(TAG,"Failed connecting to service");
-            }
-        });
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.home, menu);
+        return true;
     }
-
-    /* register the broadcast receiver with the intent values to be matched */
-    @Override
-    public void onResume() {
-        super.onResume();
-        mReceiver = new WiFiDirectBroadcastReceiver(mManager, mChannel, this,peerListListener);
-        registerReceiver(mReceiver, mIntentFilter);
-
-
-
-    }
-
-
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
@@ -175,86 +89,38 @@ public class HomeActivity extends Activity implements WifiP2pManager.ConnectionI
         return super.onOptionsItemSelected(item);
     }
 
+    /**
+     * Verify that we can enable the Button that initiate the session.
+     * by checking the EditText s of the Activity
+     */
+    private final class EditTextWatcher implements TextWatcher {
 
+        @Override
+        public void beforeTextChanged(CharSequence charSequence, int start, int count, int after) {
 
-
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        unregisterReceiver(mReceiver);
-    }
-
-    @Override
-    protected void onStop() {
-        if (mManager != null && mChannel != null) {
-            mManager.removeGroup(mChannel, new WifiP2pManager.ActionListener() {
-                @Override
-                public void onFailure(int reasonCode) {
-                    Log.d(TAG, "Disconnect failed. Reason :" + reasonCode);
-                }
-                @Override
-                public void onSuccess() {
-                }
-            });
         }
-        super.onStop();
-    }
 
-    @Override
-    public void onConnectionInfoAvailable(WifiP2pInfo p2pInfo) {
-        Thread handler = null;
-        /*
-        * The group owner accepts connections using a server socket and then spawns a
-        * client socket for every client.
-        */
-        if (p2pInfo.isGroupOwner) {
-            Log.d(TAG, "Connected as group owner");
-            try {
-                handler = new Server(
-                        ((Server.MessageTarget) this).getHandler());
-                handler.start();
-            } catch (IOException e) {
-                Log.d(TAG,
-                        "Failed to create a server thread - " + e.getMessage());
-                return;
+        @Override
+        public void onTextChanged(CharSequence charSequence, int start, int before, int count) {
+            if (chatButtonCanBeEnabled()) {
+                enableStartButton(true);
+            } else {
+                enableStartButton(false);
             }
-        } else {
-            Log.d(TAG, "Connected as peer");
-            handler = new Client(
-                    ((Server.MessageTarget) this).getHandler(),
-                    p2pInfo.groupOwnerAddress);
-            handler.start();
         }
 
-        chatRoom = new ChatRoom();
-        getFragmentManager().beginTransaction()
-                .replace(R.id.container, chatRoom).commit();
+        @Override
+        public void afterTextChanged(Editable editable) {
 
-    }
-
-    @Override
-    public Handler getHandler() {
-        return mHandler;
-    }
-
-    @Override
-    public boolean handleMessage(Message message) {
-
-        switch (message.what) {
-            case MESSAGE_READ:
-                byte[] readBuf = (byte[]) message.obj;
-                // construct a string from the valid bytes in the buffer
-                String readMessage = new String(readBuf, 0, message.arg1);
-                Log.d(TAG, readMessage);
-                (chatRoom).pushMessage("Peer: " + readMessage);
-                break;
-            case MY_HANDLE:
-                Object obj = message.obj;
-                (chatRoom).setMyConnection((MyConnection) obj);
         }
-        return true;
-    }
 
+        private boolean chatButtonCanBeEnabled() {
+            if ((((EditText) findViewById(R.id.status_message)).getText().length() >= MIN_STATUS_LENGTH)
+                    && (((EditText) findViewById(R.id.username)).getText().length() >= MIN_USERNAME_LENGTH)) {
+                return true;
+            }
+            return false;
+        }
+    }
 
 }
