+++ app/src/main/java/ch/tarsier/tarsier/WiFiDirectBroadcastReceiver.java
@@ -22,14 +22,14 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver  {
     private WifiP2pManager.Channel mChannel;
     private WifiP2pManager.PeerListListener peerListListener;
 
-    private HomeActivity mActivity;
+    private WiFiDirectDebugActivity mActivity;
     Server server = null;
 
 
 
 
     public WiFiDirectBroadcastReceiver(WifiP2pManager manager, WifiP2pManager.Channel channel,
-                                       HomeActivity activity,WifiP2pManager.PeerListListener peerListListener) {
+                                       WiFiDirectDebugActivity activity,WifiP2pManager.PeerListListener peerListListener) {
         super();
         this.mManager = manager;
         this.mChannel = channel;
@@ -40,7 +40,7 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver  {
     @Override
     public void onReceive(Context context, Intent intent) {
         String action = intent.getAction();
-        Log.d(HomeActivity.TAG, action);
+        Log.d(WiFiDirectDebugActivity.TAG, action);
         if (WifiP2pManager.WIFI_P2P_CONNECTION_CHANGED_ACTION.equals(action)) {
             if (mManager == null) {
                 return;
@@ -50,7 +50,7 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver  {
             if (networkInfo.isConnected()) {
                 // we are connected with the other device, request connection
                 // info to find group owner IP
-                Log.d(HomeActivity.TAG,
+                Log.d(WiFiDirectDebugActivity.TAG,
                         "Connected to p2p network. Requesting network details");
                 mManager.requestConnectionInfo(mChannel,
                         (WifiP2pManager.ConnectionInfoListener) mActivity);
@@ -61,7 +61,7 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver  {
                 .equals(action)) {
             WifiP2pDevice device = (WifiP2pDevice) intent
                     .getParcelableExtra(WifiP2pManager.EXTRA_WIFI_P2P_DEVICE);
-            Log.d(HomeActivity.TAG, "Device status -" + device.status);
+            Log.d(WiFiDirectDebugActivity.TAG, "Device status -" + device.status);
         } else if(WifiP2pManager.WIFI_P2P_PEERS_CHANGED_ACTION.equals(action)){
             // Request available peers from the wifi p2p manager. This is an
             // asynchronous call and the calling activity is notified with a
@@ -73,7 +73,4 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver  {
 
         }
     }
-
-
-
 }
