+++ app/src/main/res/layout/activity_add_profile_picture.xml
@@ -1,0 +1,44 @@
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:background="@drawable/home_background"
+    tools:context="ch.tarsier.tarsier.AddProfilePictureActivity">
+
+    <RelativeLayout
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:layout_centerVertical="true"
+        >
+
+
+        <Button
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/add_from_existing"
+            android:layout_centerHorizontal="true"
+            android:layout_marginBottom="@dimen/view_spacing"
+            android:text="@string/add_from_existing"
+            android:textColor="@android:color/white"
+            android:background="@drawable/tarsier_green_base"
+            android:onClick="onClickAddExisting"/>
+
+        <Button
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/add_new_picture"
+            android:text="@string/add_new_picture"
+            android:textColor="@android:color/white"
+            android:layout_below="@id/add_from_existing"
+            android:background="@drawable/tarsier_green_base"
+            android:onClick="onClickNewPicture"
+            />
+
+
+        </RelativeLayout>
+
+</RelativeLayout>
