+++ app/src/main/res/layout/activity_home.xml
@@ -1,18 +1,68 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:orientation="vertical" >
-    <LinearLayout
-        android:id="@+id/container"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        android:layout_gravity="top"
-        android:orientation="vertical" />
-    <TextView
-        android:id="@+id/status_text"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    tools:context=".HomeActivity"
+    android:background="@drawable/home_background"
+    >
+    <RelativeLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="bottom" >
-    </TextView>
-</FrameLayout>
+        android:layout_centerVertical="true"
+        >
+
+        <RelativeLayout
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/relative_layout_picture_username"
+            android:layout_marginBottom="@dimen/view_spacing"
+            >
+
+            <ImageView
+                android:layout_width="@dimen/add_picture"
+                android:layout_height="@dimen/add_picture"
+                android:id="@+id/picture"
+                android:src="@drawable/add_picture_home"
+                android:onClick="onClickAddPicture"
+                />
+
+            <EditText
+                android:layout_width="fill_parent"
+                android:layout_height="wrap_content"
+                android:hint="@string/hint_username"
+                android:id="@+id/username"
+                android:layout_toEndOf="@id/picture"
+                android:layout_centerInParent="true"
+                />
+
+        </RelativeLayout>
+
+        <EditText
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:hint="@string/hint_status_message"
+            android:id="@+id/status_message"
+            android:layout_marginBottom="@dimen/view_spacing"
+            android:layout_below="@id/relative_layout_picture_username"
+            />
+
+        <Button
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/lets_chat"
+            android:text="@string/lets_chat"
+            android:textColor="@android:color/white"
+            android:textSize="@dimen/lets_chat_size"
+            android:background="@drawable/tarsier_green_base"
+            android:layout_centerHorizontal="true"
+            android:layout_below="@id/status_message"
+            android:onClick="onClickLetsChat"
+            />
+
+    </RelativeLayout>
+
+</RelativeLayout>
