+++ app/src/main/res/values/dimens.xml
@@ -2,4 +2,8 @@
     <!-- Default screen margins, per the Android Design guidelines. -->
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
+    <dimen name="add_picture">100dp</dimen>
+    <dimen name="view_spacing">20dp</dimen>
+
+    <dimen name="lets_chat_size">14pt</dimen>
 </resources>
