+++ app/src/main/res/values/strings.xml
@@ -7,5 +7,11 @@
     <string name="btn_send">Send Message</string>
     <string name="finding_service">Finding local peers</string>
     <string name="action_settings">Settings</string>
+    <string name="lets_chat">Let\'s chat!</string>
+    <string name="hint_status_message">Status message</string>
+    <string name="hint_username">Username</string>
+    <string name="title_activity_add_profile_picture">Add a profile picture</string>
+    <string name="add_from_existing">Add from your gallery</string>
+    <string name="add_new_picture">Take a new picture</string>
 
 </resources>
